refactor: PaymentResult, PaymentConfirmation 패키지 조정
