+++ src/main/java/roomescape/controller/PaymentController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.payment.PaymentResult;
+import roomescape.domain.PaymentResult;
 import roomescape.infrastructure.LoginRequired;
 import roomescape.service.PaymentService;
