+++ src/main/java/roomescape/domain/PaymentGateway.java
@@ -1,8 +1,5 @@
 package roomescape.domain;
 
-import roomescape.dto.payment.PaymentConfirmation;
-import roomescape.dto.payment.PaymentResult;
-
 public interface PaymentGateway {
 
     PaymentResult confirm(PaymentConfirmation confirmation);
