+++ src/main/java/roomescape/domain/PaymentResult.java
@@ -1,6 +1,4 @@
-package roomescape.dto.payment;
-
-import roomescape.domain.PaymentStatus;
+package roomescape.domain;
 
 public record PaymentResult(
         String paymentKey,
