+++ src/main/java/roomescape/infrastructure/payment/TossPaymentGateway.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
 import roomescape.domain.PaymentGateway;
 import roomescape.domain.PaymentStatus;
-import roomescape.dto.payment.PaymentConfirmation;
-import roomescape.dto.payment.PaymentResult;
+import roomescape.domain.PaymentConfirmation;
+import roomescape.domain.PaymentResult;
 import roomescape.infrastructure.payment.dto.ConfirmRequest;
 import roomescape.infrastructure.payment.dto.TossErrorResponse;
 import roomescape.infrastructure.payment.dto.TossPaymentResponse;
