+++ src/main/java/roomescape/service/PaymentService.java
@@ -7,8 +7,8 @@ import roomescape.domain.OrderId;
 import roomescape.domain.PaymentGateway;
 import roomescape.domain.PaymentStatus;
 import roomescape.domain.ReservationStatus;
-import roomescape.dto.payment.PaymentConfirmation;
-import roomescape.dto.payment.PaymentResult;
+import roomescape.domain.PaymentConfirmation;
+import roomescape.domain.PaymentResult;
 import roomescape.exception.ErrorType;
 import roomescape.exception.RoomescapeException;
 import roomescape.repository.OrderRepository;
