+++ src/test/java/roomescape/service/PaymentServiceTest.java
@@ -13,8 +13,8 @@ import roomescape.domain.Order;
 import roomescape.domain.PaymentGateway;
 import roomescape.domain.PaymentStatus;
 import roomescape.domain.Reservation;
-import roomescape.dto.payment.PaymentConfirmation;
-import roomescape.dto.payment.PaymentResult;
+import roomescape.domain.PaymentConfirmation;
+import roomescape.domain.PaymentResult;
 import roomescape.exception.RoomescapeException;
 import roomescape.fixture.DbFixtures;
 import roomescape.fixture.Fixtures;
